+++ src/worldview_collider_tool.cpp
@@ -119,18 +119,18 @@ WorldViewColliderTool::on_primary_button_release (int x, int y)
 void
 WorldViewColliderTool::on_mouse_move (int x, int y, int of_x, int of_y)
 {
-  Vector2d current_pos = WorldViewComponent::instance()->get_gc()->screen_to_world(Vector2d(x,y));
+  Vector2d current_pos = WorldViewComponent::instance()->get_gc()->screen_to_world(Vector2d(static_cast<float>(x), static_cast<float>(y)));
 
   if (m_move_collider)
-    {
-      m_move_collider->set_pos(current_pos - m_click_pos);
-    }
+  {
+    m_move_collider->set_pos(current_pos - m_click_pos);
+  }
 }
 
 void
 WorldViewColliderTool::on_secondary_button_press (int x, int y)
 {
-  m_to_delete_collider = get_collider(WorldViewComponent::instance()->get_gc()->screen_to_world(Vector2d(x, y)));
+  m_to_delete_collider = get_collider(WorldViewComponent::instance()->get_gc()->screen_to_world(Vector2d(static_cast<float>(x), static_cast<float>(y))));
 }
 
 void
@@ -139,7 +139,7 @@ WorldViewColliderTool::on_secondary_button_release (int x, int y)
   World& world = *Controller::instance()->get_world();
 
   if (m_to_delete_collider
-      == get_collider(WorldViewComponent::instance()->get_gc()->screen_to_world(Vector2d(x, y))))
+      == get_collider(WorldViewComponent::instance()->get_gc()->screen_to_world(Vector2d(static_cast<float>(x), static_cast<float>(y)))))
     {
       Controller::instance()->push_undo();
       world.remove_collider(m_to_delete_collider);
